requirements.txt
python-magic-0.4.12
ipfs-api-0.2.3
